# Avoid double serialization of message data