--- a/activesupport/lib/active_support/messages/metadata.rb
+++ b/activesupport/lib/active_support/messages/metadata.rb
@@ -1,82 +1,100 @@
 # frozen_string_literal: true
 
 require "time"
+require "active_support/json"
 
 module ActiveSupport
   module Messages # :nodoc:
-    class Metadata # :nodoc:
-      def initialize(message, expires_at = nil, purpose = nil)
-        @message, @purpose = message, purpose
-        @expires_at = expires_at.is_a?(String) ? parse_expires_at(expires_at) : expires_at
-      end
+    module Metadata # :nodoc:
+      singleton_class.attr_accessor :use_message_serializer_for_metadata
 
-      def as_json(options = {})
-        { _rails: { message: @message, exp: @expires_at, pur: @purpose } }
-      end
-
-      class << self
-        def wrap(message, expires_at: nil, expires_in: nil, purpose: nil)
-          if expires_at || expires_in || purpose
-            JSON.encode new(encode(message), pick_expiry(expires_at, expires_in), purpose)
-          else
-            message
-          end
-        end
-
-        def verify(message, purpose)
-          extract_metadata(message).verify(purpose)
-        end
-
-        private
-          def pick_expiry(expires_at, expires_in)
-            if expires_at
-              expires_at.utc.iso8601(3)
-            elsif expires_in
-              Time.now.utc.advance(seconds: expires_in).iso8601(3)
-            end
-          end
-
-          def extract_metadata(message)
-            begin
-              data = JSON.decode(message) if message.start_with?('{"_rails":')
-            rescue ::JSON::JSONError
-            end
-
-            if data
-              new(decode(data["_rails"]["message"]), data["_rails"]["exp"], data["_rails"]["pur"])
-            else
-              new(message)
-            end
-          end
-
-          def encode(message)
-            ::Base64.strict_encode64(message)
-          end
-
-          def decode(message)
-            ::Base64.strict_decode64(message)
-          end
-      end
-
-      def verify(purpose)
-        @message if match?(purpose) && fresh?
-      end
+      ENVELOPE_SERIALIZERS = [
+        ::JSON,
+        ActiveSupport::JSON,
+        ActiveSupport::JsonWithMarshalFallback,
+        Marshal,
+      ]
 
       private
-        def match?(purpose)
-          @purpose.to_s == purpose.to_s
+        def serialize_with_metadata(data, **metadata)
+          has_metadata = metadata.any? { |k, v| v }
+
+          if has_metadata && !use_message_serializer_for_metadata?
+            data_string = serialize_to_json_safe_string(data)
+            envelope = wrap_in_metadata_envelope({ "message" => data_string }, **metadata)
+            ActiveSupport::JSON.encode(envelope)
+          else
+            data = wrap_in_metadata_envelope({ "data" => data }, **metadata) if has_metadata
+            serializer.dump(data)
+          end
         end
 
-        def fresh?
-          @expires_at.nil? || Time.now.utc < @expires_at
+        def deserialize_with_metadata(message, **expected_metadata)
+          if dual_serialized_metadata_envelope_json?(message)
+            envelope = ActiveSupport::JSON.decode(message)
+            extracted = extract_from_metadata_envelope(envelope, **expected_metadata)
+            deserialize_from_json_safe_string(extracted["message"]) if extracted
+          else
+            deserialized = serializer.load(message)
+            if metadata_envelope?(deserialized)
+              extracted = extract_from_metadata_envelope(deserialized, **expected_metadata)
+              extracted["data"] if extracted
+            else
+              deserialized if expected_metadata.none? { |k, v| v }
+            end
+          end
         end
 
-        def parse_expires_at(expires_at)
-          if ActiveSupport.use_standard_json_time_format
+        def use_message_serializer_for_metadata?
+          Metadata.use_message_serializer_for_metadata && Metadata::ENVELOPE_SERIALIZERS.include?(serializer)
+        end
+
+        def wrap_in_metadata_envelope(hash, expires_at: nil, expires_in: nil, purpose: nil)
+          expiry = pick_expiry(expires_at, expires_in)
+          hash["exp"] = expiry if expiry
+          hash["pur"] = purpose.to_s if purpose
+          { "_rails" => hash }
+        end
+
+        def extract_from_metadata_envelope(envelope, purpose: nil)
+          hash = envelope["_rails"]
+          return if hash["exp"] && Time.now.utc >= parse_expiry(hash["exp"])
+          return if hash["pur"] != purpose&.to_s
+          hash
+        end
+
+        def metadata_envelope?(object)
+          object.is_a?(Hash) && object.key?("_rails")
+        end
+
+        def dual_serialized_metadata_envelope_json?(string)
+          string.start_with?('{"_rails":{"message":')
+        end
+
+        def pick_expiry(expires_at, expires_in)
+          if expires_at
+            expires_at.utc.iso8601(3)
+          elsif expires_in
+            Time.now.utc.advance(seconds: expires_in).iso8601(3)
+          end
+        end
+
+        def parse_expiry(expires_at)
+          if !expires_at.is_a?(String)
+            expires_at
+          elsif ActiveSupport.use_standard_json_time_format
             Time.iso8601(expires_at)
           else
             Time.parse(expires_at)
           end
+        end
+
+        def serialize_to_json_safe_string(data)
+          ::Base64.strict_encode64(serializer.dump(data))
+        end
+
+        def deserialize_from_json_safe_string(string)
+          serializer.load(::Base64.strict_decode64(string))
         end
     end
   end